+++ BE-Byundohwi/src/main/java/org/example/bebyundohwi/week1/global/exception/security/auth/AuthDetailsService.java
@@ -1,4 +1,4 @@
-package org.example.bebyundohwi.week1.global.security.auth;
+package org.example.bebyundohwi.week1.global.exception.security.auth;
 
 import lombok.RequiredArgsConstructor;
 import org.example.bebyundohwi.week1.domain.user.domain.UserEntity;
