+++ BE-Byundohwi/src/main/java/org/example/bebyundohwi/week1/global/exception/security/exception/ExpiredTokenException.java
@@ -1,0 +1,12 @@
+package org.example.bebyundohwi.week1.global.exception.security.exception;
+
+import org.example.bebyundohwi.week1.global.error.ErrorCode;
+import org.example.bebyundohwi.week1.global.exception.CustomException;
+
+public class ExpiredTokenException extends CustomException {
+    public static final CustomException EXCEPTION = new ExpiredTokenException();
+
+    public ExpiredTokenException() {
+        super(ErrorCode.EXPIRED_TOKEN);
+    }
+}
