+++ BE-Byundohwi/src/main/java/org/example/bebyundohwi/week1/global/exception/security/role/Role.java
@@ -1,4 +1,4 @@
-package org.example.bebyundohwi.week1.global.security.role;
+package org.example.bebyundohwi.week1.global.exception.security.role;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
