--- BE-Byundohwi/src/main/java/org/example/bebyundohwi/week1/global/security/auth/AuthDetails.java
@@ -1,54 +1,0 @@
-package org.example.bebyundohwi.week1.global.security.auth;
-
-import lombok.AllArgsConstructor;
-import org.example.bebyundohwi.week1.domain.user.domain.UserEntity;
-import org.example.bebyundohwi.week1.global.security.role.Role;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-@AllArgsConstructor
-public class AuthDetails implements UserDetails {
-    private final UserEntity user;
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return UserDetails.super.isAccountNonExpired();
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return UserDetails.super.isAccountNonLocked();
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return UserDetails.super.isCredentialsNonExpired();
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return UserDetails.super.isEnabled();
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        List<GrantedAuthority> authorities = new ArrayList<>();
-        authorities.add(new SimpleGrantedAuthority("ROLE_"+user.getRole().toString()));
-        return authorities;
-    }
-
-    @Override
-    public String getPassword() {
-        return null;
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getUsername();
-    }
-}
